% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.alistfirm-package.R
\docType{package}
\name{db.alistfirm-package}
\alias{db.alistfirm}
\alias{db.alistfirm-package}
\title{db.alistfirm: Database of Chinese listed firms}
\description{
Database of Chinese A-shared listed firms, including variables of basic information, finance status, corporate governance, etc.
}
\keyword{internal}
