% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_alistfirm}
\alias{db_alistfirm}
\title{A-share listed firm}
\format{
It has 49 variables:
\describe{
  \item{stkcd}{stock code, int}
  \item{year}{year, int}
  \item{province}{office address: province, chr}
  \item{city}{office address: city, chr}
  \item{ind_code}{industry code CSRC-2012, chr}
}
}
\source{
CSMAR, Wind
}
\description{
A dataset containing listed firm data
    during 2001 ~ 2021.
}
