% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_alistfirm}
\alias{db_alistfirm}
\title{A-share listed firm}
\format{
It has 49 variables:
\describe{
  \item{stkcd}{stock code}
  \item{year}{year}
  \item{province}{office address: province}
  \item{prov_cs}{cross-sectional province, from CSMAR}
  \item{city}{office address: city}
  \item{city_cs}{cross-sectional city, from CSMAR}
  \item{ind_code}{industry code CSRC-2012}
  \item{ind_cs}{cross-sectional industry, from CSMAR}
  \item{ind_class}{industry class, only six categories}
  \item{admin_expense}{administration (handling) expense}
  \item{book_equity}{book value of equity}
  \item{btm}{book-to-market ratio}
  \item{cash}{cash holdings}
  \item{cash_ratio}{cash holdings over total assets}
  \item{ceo_duality}{whether CEO and general manager is the same}
  \item{ceo_duality_num}{numerical version of `ceo_duality`}
  \item{checkandbalance_equity}{(`top10_share` - `top1_share`) / `top1_share`}
  \item{citation}{total number of patent citations}
  \item{construction}{net amount of construction in process}
  \item{current_asset}{current assets}
  \item{current_asset_to}{current assets turnover}
  \item{current_debt}{current liabilities}
  \item{digit_a}{digitalization based on text analysis, see 10.19744/j.cnki.11-1235/f.2021.0097}
  \item{digit_b}{digitalization based on text analysis, see 10.19795/j.cnki.cn11-1166/f.20210705.001}
  \item{director_size}{number of directors}
  \item{director_wage}{total salary of directors}
  \item{dividend_pay}{dividend payment}
  \item{ebitda}{EBITDA}
  \item{fix_asset}{fixed assets}
  \item{inddir_absence}{1 if any independent director has attended board meeting, 0 otherwise}
  \item{inddir_rate}{number of independent directors over `director_size`}
  \item{inddir_workplace}{whether the office of independent director located in the same city as the firms}
  \item{intan_ratio}{intangible assets over total assets}
  \item{invest_pay}{investment payment}
  \item{invest_receive}{investment receivings}
  \item{invisible_asset}{invisible assets}
  \item{labor}{number of staff}
  \item{length_all}{total word count of annual report}
  \item{length_txt}{word count of Chinese/English in annual report}
  \item{long_borrow}{long-term loan}
  \item{long_pay}{long-term payable}
  \item{market_equity}{market value of equity}
  \item{net_profit}{net profit}
  \item{operate_cash}{operating cash flow}
  \item{patent}{patent application}
  \item{patent_grt}{patent application that granted}
  \item{patent_inv}{patent application for invention that granted}
  \item{rd}{R&D}
  \item{rd_expense}{expensed R&D}
  \item{rd_invest}{capitalized R&D}
  \item{rd_person}{research staff}
  \item{rd_person_ratio}{`rd_person` / `labor`}
  \item{receive_ratio}{receivable assets over total assets}
  \item{retain_earn}{retained earnings}
  \item{return}{yearly averaged stock return}
  \item{revenue}{operating revenue}
  \item{roa}{ROA}
  \item{roe}{ROE}
  \item{sale_expense}{selling expense}
  \item{seperation}{seperation of control right and ownership right}
  \item{short_borrow}{short-term loan}
  \item{soe_ccer}{whether state-owned-enterprise (SOE), from CCER}
  \item{soe_ccer_num}{numerical version of `soe_ccer`}
  \item{soe_csmar}{whether SOE, from CSMAR}
  \item{soe_csmar_num}{numerical version of `soe_csmar`}
  \item{special_pay}{special payment}
  \item{st}{is this firm coded "ST"?}
  \item{state}{ownership type, from CSMAR}
  \item{state_cs}{cross-sectional ownershiptype, from CSMAR}
  \item{supervisor_size}{number of supervisors}
  \item{tobinq}{Tobin's Q ratio}
  \item{top10_share}{share prop. of the largest ten shareholder}
  \item{top1_share}{share proportion of the largest shareholder}
  \item{total_asset}{total assets}
  \item{total_debt}{total liabilities}
  \item{year_estb}{the year of establishment}
  \item{year_list}{the year of IPO}
}
}
\source{
CSMAR, Wind, CCER
}
\description{
A dataset containing listed firm data during 2001 ~ 2021.
    For detailed Chinese information, see
    https://github.com/yang-yuchuan/db.alistfirm/blob/master/R/data.R
}
